+++ src/additionalScripts/autoSaveAssignment/index.ts
