+++ src/index.ts
@@ -1,7 +1,7 @@
 import './custom.css';
 import '../public/manifest.json';
 import '../public/background';
-import './additionalScripts/autoSaveAssignment/customSubmitAssignmentForm';
+import './additionalScripts/autoSaveAssignment';
 import '../assets/digicamExtensionLogo128.png';
 import '../assets/digicamExtensionLogo48.png';
 import '../assets/digicamExtensionLogo16.png';
